proxy-server/requirements.txt
kafka-python~=2.0.0
flask~=2.1.0